Add files via upload
This is a project of Vivek Venigalla made around 3 years ago. This has some functionalities that our web app may need. If you want to use this application, you will have to download MAMP(creates a local host on the computer) and host the website. I will attach screenshots of the product if you do not want to run it.